backend/requirements.txt
# Framework web
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
Flask-Cors==4.0.0

# Authentification et sécurité
Flask-JWT-Extended==4.5.3
bcrypt==4.1.1

# Utilitaires
python-dotenv==1.0.0
pytz==2023.3

# Dépendances de base
Werkzeug==2.3.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
SQLAlchemy==2.0.21
aniso8601==9.0.1
six==1.16.0


